% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_tg.R
\name{.coerce_df}
\alias{.coerce_df}
\title{.coerce_df internal}
\usage{
.coerce_df(x, vec)
}
\arguments{
\item{x}{A data.frame object with columns that need to be coerced}

\item{vec}{A reference data.frame object with target data types.}
}
\value{
A data.frame object with the same structure as \code{x}, but with columns
coerced to match those in \code{vec}.
}
\description{
.coerce_df internal
}
\keyword{internal}
