% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg_normalize.r
\name{tg_normalize}
\alias{tg_normalize}
\title{Validate Telegram Data}
\usage{
tg_normalize(tg_list, lib = "telescrape", bind = FALSE, datecol = FALSE)
}
\arguments{
\item{tg_list}{A list of data.frame objects or a data.frame object. All data
in each data.frame object have their vector types normalized.}

\item{lib}{A string with the name of the library API that constructed the
supplied data.frame object.}

\item{bind}{A Boolean argument to determine if the list of data.frame objects
should be bound into a single data.frame object. If TRUE, a single data.frame
is returned. If FALSE, a list of data.frames is returned.}

\item{datecol}{A Boolean argument to add a "date" column, which is an as.date()
coercion that makes per day statistical data easier.}
}
\value{
A data.frame (or list of data.frames if bind=FALSE)
}
\description{
\code{tg_normalize} normalizes the input data from a list of data.frames or a data.frame
object with the option of binding them to a single data.frame.
}
