% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_tg.R
\name{validate_tg}
\alias{validate_tg}
\title{Processing data returned from the telethon channel.getMessages Python Telethon
interface for accessing the Telegram API}
\usage{
validate_tg(tg_list, bind = FALSE)
}
\arguments{
\item{tg_list}{a list of data.frame objects or a data.frame object. All data
in each data.frame object have their vector types normalized}

\item{bind}{a boolean arg to coerce to data.frame object}
}
\description{
Processing data returned from the telethon channel.getMessages Python Telethon
interface for accessing the Telegram API
}
